[basic-components-sample]: rename dialog-field-test to dialog-field-sample
Renamed all classes and content paths from *Test* to *Sample* to avoid
confusion with test classes. Package renamed from dialogfieldtest to
dialogfieldsample. Resource type updated to dialog-field-sample.
