+++ basic-components-sample/src/content/jcr_root/libs/kestros/samples/components/content/dialog-field-sample.json
@@ -1,6 +1,6 @@
 {
   "jcr:primaryType": "kes:ComponentType",
-  "jcr:title": "Dialog Field Test",
+  "jcr:title": "Dialog Field Sample",
   "jcr:description": "Sample component that exercises every available KDF dialog field type. Renders authored values in a debug/diagnostic table.",
   "componentGroup": "Sample Components",
   "sling:resourceSuperType": "kestros/commons/components/kestros-parent",
