+++ basic-components-sample/src/content/jcr_root/libs/kestros/samples/components/content/dialog-field-sample/common/content.html
@@ -1,7 +1,7 @@
-<sly data-sly-use.component="${'io.kestros.samples.content.dialogfieldtest.DialogFieldTestDataSourceComponent'}"/>
+<sly data-sly-use.component="${'io.kestros.samples.content.dialogfieldsample.DialogFieldSampleDataSourceComponent'}"/>
 
-<div class="dialog-field-test">
-    <h3>Dialog Field Test - Debug Output</h3>
+<div class="dialog-field-sample">
+    <h3>Dialog Field Sample - Debug Output</h3>
     <table border="1" cellpadding="4" cellspacing="0" style="border-collapse: collapse; width: 100%;">
         <thead>
             <tr>
