+++ basic-components-sample/src/content/jcr_root/libs/kestros/samples/components/content/dialog-field-sample/datasources/default.json
@@ -2,7 +2,7 @@
   "jcr:primaryType": "nt:unstructured",
   "jcr:title": "Static/Default",
   "group": "Sample Components",
-  "classPath": "io.kestros.samples.content.dialogfieldtest.DialogFieldTestStaticDataSource",
+  "classPath": "io.kestros.samples.content.dialogfieldsample.DialogFieldSampleStaticDataSource",
   "edit": {
     "jcr:primaryType": "nt:unstructured",
     "sling:resourceType": "kestros/cms2/dialog",
